Split ClrPrepare where Calcite splits Prepare, and fix the docs the moves left
The prepare pipeline was one class. Calcite has two, and I collapsed them on the
argument that Prepare has exactly one subclass — an observation about what exists
rather than a decision, which is the same reasoning that left PrepareRel out.

The split it makes is real. Prepare holds the algorithm and cannot reach a
CalciteSchema, a RelOptCluster or a SparkHandler; CalcitePreparingStmt holds that
wiring. It is where a preparing statement backed by something other than a
CalcitePrepare.Context would part company, and it keeps the port diffable: when
Prepare.prepareSql changes upstream the change lands in one file rather than
having to be found inside a larger one.

So ClrPrepare is now the mirror of Prepare, ClrPreparingStmt the mirror of
CalcitePreparingStmt, and the driver is renamed ClrPrepareImpl after the class it
actually mirrors. The split surfaced a second misplacement: mapTableModOp is
protected on Prepare and I had it as a private static on
ClrEnumerablePreparingStmt, two levels down.

One thing the split cannot mirror. Calcite's driver reads getSqlValidator though
it is protected, because CalcitePreparingStmt is nested inside CalcitePrepareImpl
and Java lets an outer class reach a nested one"s protected members. C# has no
equivalent, so there is a public accessor beside the protected member rather than
the member"s protection being widened.

Doc comments audited against what the code does now. Five described machinery
that no longer exists — a CalciteSignature the session no longer sees, a
PrepareFactory that is gone, a Bindable that is now an IClrBindable — and one
asserted something Calcite does not do: that Calcite holds an uncompiled plan
where we compile first. It does not. compileToBindable cooks the source, loads
the class and calls newInstance, so both compile at the same point and only the
artifact differs.

The Extensions README described the package as connection-property helpers, which
stopped being true when the conventions moved in, and its sample imported the old
namespace. Rewritten for the audience that reads it on NuGet.

IDE0008 is suppressed in Directory.Build.props. Twenty-nine stale cref prefixes
left by flattening the convention are fixed, and the build is free of CS15xx.
