+++ src/Apache.Calcite.Extensions/Prepare/ClrPreparingStmt.cs
@@ -11,38 +11,28 @@ using org.apache.calcite.sql;
 using org.apache.calcite.sql.parser;
 using org.apache.calcite.sql.validate;
 using org.apache.calcite.sql2rel;
-using org.apache.calcite.tools;
-using org.apache.calcite.util;
 
 namespace Apache.Calcite.Extensions.Prepare
 {
 
     /// <summary>
-    /// Takes a parsed statement to a plan, and hands the plan to a convention to implement.
+    /// A statement being prepared against a Calcite schema.
     /// </summary>
     /// <remarks>
-    /// Our counterpart of <c>Prepare</c> and <c>CalcitePrepareImpl.CalcitePreparingStmt</c>, which are one
-    /// class here because <c>Prepare</c>'s only subclass in Calcite is that one. The steps are its
-    /// <c>prepareSql</c>, in order: convert, checked arithmetic, flatten, decorrelate, trim, optimize,
-    /// implement — with the two <c>EXPLAIN</c> exits where Calcite has them.
+    /// Our counterpart of <c>CalcitePrepareImpl.CalcitePreparingStmt</c>: the wiring half, holding the
+    /// cluster, the planner, the convertlet table and the schema, and answering everything
+    /// <see cref="ClrPrepare"/> leaves abstract. It is also the view expander a
+    /// <c>SqlToRelConverter</c> is given.
     ///
-    /// <para>It is written rather than derived because <c>Prepare.implement</c> is declared to return a
-    /// <c>PreparedResult</c>, an interface whose reason for existing is <c>getBindable</c> — a linq4j
-    /// <c>Enumerable</c>. Our conventions compile to a delegate, so there is nothing to hand back through
-    /// it. <see cref="ClrPrepareResult"/> is that interface without the member.</para>
-    ///
-    /// <para>Three things are a convention's own and nothing else is: the convention a plan must end in,
-    /// the program that gets it there, and the compiler at the end. They are the three abstract members
-    /// below, and they are the whole of what a second convention supplies.</para>
+    /// <para>Three things are still a convention's own and nothing else is: the convention a plan must end
+    /// in, the program that gets it there, and the compiler at the end. A subclass supplies those, and that
+    /// is the whole of what a second convention writes.</para>
     /// </remarks>
-    abstract class ClrPreparingStmt : RelOptTable.ViewExpander
+    abstract class ClrPreparingStmt : ClrPrepare, RelOptTable.ViewExpander
     {
 
-        readonly CalcitePrepare.Context context;
-        readonly CalciteCatalogReader catalogReader;
         readonly CalciteSchema schema;
         readonly RelOptCluster cluster;
-        readonly RelOptPlanner planner;
         readonly RexBuilder rexBuilder;
         readonly JavaTypeFactory typeFactory;
         readonly SqlRexConvertletTable convertletTable;
@@ -58,8 +48,6 @@ namespace Apache.Calcite.Extensions.Prepare
         readonly java.util.Map internalParameters = new java.util.LinkedHashMap();
 
         SqlValidator? validator;
-        RelDataType? parameterRowType;
-        java.util.List? fieldOrigins;
 
         /// <summary>
         /// Initializes a new instance.
@@ -69,23 +57,17 @@ namespace Apache.Calcite.Extensions.Prepare
             CalciteCatalogReader catalogReader,
             CalciteSchema schema,
             RelOptCluster cluster,
-            SqlRexConvertletTable convertletTable)
+            Convention resultConvention,
+            SqlRexConvertletTable convertletTable) :
+            base(context, catalogReader, resultConvention)
         {
-            this.context = context ?? throw new ArgumentNullException(nameof(context));
-            this.catalogReader = catalogReader ?? throw new ArgumentNullException(nameof(catalogReader));
             this.schema = schema ?? throw new ArgumentNullException(nameof(schema));
             this.cluster = cluster ?? throw new ArgumentNullException(nameof(cluster));
             this.convertletTable = convertletTable ?? throw new ArgumentNullException(nameof(convertletTable));
 
-            planner = cluster.getPlanner();
             rexBuilder = cluster.getRexBuilder();
             typeFactory = (JavaTypeFactory)cluster.getTypeFactory();
         }
-
-        /// <summary>
-        /// Gets the context the statement is prepared against.
-        /// </summary>
-        protected CalcitePrepare.Context Context => context;
 
         /// <summary>
         /// Gets the cluster the plan is built in.
@@ -97,124 +79,23 @@ namespace Apache.Calcite.Extensions.Prepare
         /// </summary>
         public java.util.Map InternalParameters => internalParameters;
 
-        /// <summary>
-        /// Gets the validator the statement is validated with.
-        /// </summary>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>CalcitePreparingStmt.getSqlValidator</c>, which caches so that one statement cannot end up
         /// with two.
         /// </remarks>
-        public SqlValidator Validator => validator ??= CreateSqlValidator(catalogReader, c => c);
+        protected override SqlValidator Validator => validator ??= CreateSqlValidator(CatalogReader, c => c);
 
         /// <summary>
-        /// Gets the convention a plan must end in.
+        /// Gets the validator the statement was validated with, for the driver.
         /// </summary>
-        protected abstract Convention ResultConvention { get; }
-
-        /// <summary>
-        /// Returns the program that takes a logical plan to that convention.
-        /// </summary>
-        protected abstract Program GetProgram();
-
-        /// <summary>
-        /// Returns the traits the root of the plan must satisfy.
-        /// </summary>
-        protected abstract RelTraitSet GetDesiredRootTraitSet(RelRoot root);
-
-        /// <summary>
-        /// Compiles the chosen plan.
-        /// </summary>
-        /// <param name="root">The root of the plan, which is of <see cref="ResultConvention"/>.</param>
-        /// <returns>The compiled statement.</returns>
-        protected abstract ClrPrepareResult Implement(RelRoot root);
-
-        /// <summary>
-        /// Prepares a parsed statement.
-        /// </summary>
-        /// <param name="sqlQuery">The statement, which an <c>EXPLAIN</c> is unwrapped from.</param>
-        /// <param name="sqlNodeOriginal">The statement as parsed, before that unwrapping.</param>
-        /// <param name="needsValidation">Whether the statement still has to be validated.</param>
-        /// <returns>The compiled statement, or the rendered plan where it was an <c>EXPLAIN</c>.</returns>
-        public ClrPrepareResult PrepareSql(SqlNode sqlQuery, SqlNode sqlNodeOriginal, bool needsValidation)
-        {
-            ArgumentNullException.ThrowIfNull(sqlQuery);
-            ArgumentNullException.ThrowIfNull(sqlNodeOriginal);
-
-            var config = SqlToRelConverter.config()
-                .withTrimUnusedFields(true)
-                .withExpand(((java.lang.Boolean)org.apache.calcite.prepare.Prepare.THREAD_EXPAND.get()).booleanValue())
-                .withInSubQueryThreshold(((java.lang.Integer)org.apache.calcite.prepare.Prepare.THREAD_INSUBQUERY_THRESHOLD.get()).intValue())
-                .withExplain(sqlQuery.getKind() == SqlKind.EXPLAIN);
-
-            var configHolder = Holder.of(config);
-            org.apache.calcite.runtime.Hook.SQL2REL_CONVERTER_CONFIG_BUILDER.run(configHolder);
-
-            var sqlToRelConverter = GetSqlToRelConverter(Validator, catalogReader, (SqlToRelConverter.Config)configHolder.get());
-
-            SqlExplain? sqlExplain = null;
-            if (sqlQuery.getKind() == SqlKind.EXPLAIN)
-            {
-                // dig out the underlying SQL statement
-                sqlExplain = (SqlExplain)sqlQuery;
-                sqlQuery = sqlExplain.getExplicandum();
-                sqlToRelConverter.setDynamicParamCountInExplain(sqlExplain.getDynamicParamCount());
-            }
-
-            var root = sqlToRelConverter.convertQuery(sqlQuery, needsValidation, true);
-
-            // all arithmetic on exact types where the conformance asks for it, and all arithmetic producing
-            // an INTERVAL whatever the conformance says
-            var convertToChecked = context.config().conformance().checkedArithmetic();
-            var checkedConv = new ConvertToChecked(root.rel.getCluster().getRexBuilder(), convertToChecked);
-            root = root.withRel(checkedConv.visit(root.rel));
-            org.apache.calcite.runtime.Hook.CONVERTED.run(root.rel);
-
-            var resultType = Validator.getValidatedNodeType(sqlQuery);
-            fieldOrigins = Validator.getFieldOrigins(sqlQuery);
-            parameterRowType = Validator.getParameterRowType(sqlQuery);
-
-            // the logical plan, before view expansion, physical storage and decorrelation
-            if (sqlExplain != null)
-            {
-                switch (sqlExplain.getDepth().name())
-                {
-                    case nameof(SqlExplain.Depth.TYPE):
-                        return CreatePreparedExplanation(resultType, parameterRowType, null, sqlExplain.getFormat(), sqlExplain.getDetailLevel());
-                    case nameof(SqlExplain.Depth.LOGICAL):
-                        return CreatePreparedExplanation(null, parameterRowType, root, sqlExplain.getFormat(), sqlExplain.getDetailLevel());
-                }
-            }
-
-            root = root.withRel(FlattenTypes(root.rel, true));
-
-            // TopDownGeneralDecorrelator cannot run until the sub-queries are gone
-            if (context.config().forceDecorrelate() && context.config().topDownGeneralDecorrelationEnabled() == false)
-                root = root.withRel(Decorrelate(sqlToRelConverter, sqlQuery, root.rel));
-
-            if (((SqlToRelConverter.Config)configHolder.get()).isTrimUnusedFields())
-            {
-                root = TrimUnusedFields(root);
-                org.apache.calcite.runtime.Hook.TRIMMED.run(root.rel);
-            }
-
-            // the physical plan, after decorrelation
-            if (sqlExplain != null)
-            {
-                root = Optimize(root);
-                return CreatePreparedExplanation(null, parameterRowType, root, sqlExplain.getFormat(), sqlExplain.getDetailLevel());
-            }
-
-            root = Optimize(root);
-
-            // a DML rewritten to other DML — UPDATE to MERGE — keeps the rewrite's kind; anything else —
-            // CALL to SELECT — keeps the kind it was parsed as
-            if (root.kind.belongsTo(SqlKind.DML) == false)
-                root = root.withKind(sqlNodeOriginal.getKind());
-
-            org.apache.calcite.runtime.Hook.PLAN_BEFORE_IMPLEMENTATION.run(root);
-
-            return Implement(root);
-        }
+        /// <remarks>
+        /// Calcite's driver reads <c>getSqlValidator</c> even though it is protected, because
+        /// <c>CalcitePreparingStmt</c> is a nested class of <c>CalcitePrepareImpl</c> and Java lets an outer
+        /// class reach a nested one's protected members. C# has no equivalent, so the accessor is public
+        /// rather than the member's protection being widened.
+        /// </remarks>
+        public SqlValidator SqlValidator => Validator;
 
         /// <summary>
         /// Prepares a plan that was built rather than parsed.
@@ -225,8 +106,8 @@ namespace Apache.Calcite.Extensions.Prepare
         /// <remarks>
         /// <c>CalcitePreparingStmt.prepare_(Supplier&lt;RelNode&gt;, RelDataType)</c>, which is how Calcite
         /// answers <c>RelRunner</c> and a linq4j <c>Queryable</c>. It is a shorter pipeline than
-        /// <see cref="PrepareSql"/> and says so: there is nothing to validate, so there are no dynamic
-        /// parameters and no field origins, and no <c>EXPLAIN</c>, checked-arithmetic or decorrelation pass.
+        /// <c>PrepareSql</c> and says so: there is nothing to validate, so there are no dynamic parameters
+        /// and no field origins, and no <c>EXPLAIN</c>, checked-arithmetic or decorrelation pass.
         ///
         /// <para>Calcite reaches this through <c>Query.of(rel)</c> and <c>Query.of(queryable)</c>. The
         /// queryable half cannot be ported — <c>LixToRelTranslator</c> is package-private, takes a
@@ -252,8 +133,8 @@ namespace Apache.Calcite.Extensions.Prepare
             // no validation happened, so there is nothing to say about where a field came from or what
             // parameters the statement takes — which is what Calcite records here too
             var jdbcType = MakeStruct(rexBuilder.getTypeFactory(), resultType);
-            fieldOrigins = java.util.Collections.nCopies(jdbcType.getFieldCount(), null);
-            parameterRowType = rexBuilder.getTypeFactory().builder().build();
+            FieldOrigins = java.util.Collections.nCopies(jdbcType.getFieldCount(), null);
+            ParameterRowType = rexBuilder.getTypeFactory().builder().build();
 
             root = root.withRel(FlattenTypes(root.rel, true));
             root = TrimUnusedFields(root);
@@ -275,43 +156,7 @@ namespace Apache.Calcite.Extensions.Prepare
             return type.isStruct() ? type : typeFactory.builder().add("$0", type).build();
         }
 
-        /// <summary>
-        /// Gets the row type of the statement's dynamic parameters, once it has been validated.
-        /// </summary>
-        protected RelDataType ParameterRowType =>
-            parameterRowType ?? throw new InvalidOperationException("The statement has not been validated.");
-
-        /// <summary>
-        /// Gets, per field, the name it originates from, once the statement has been validated.
-        /// </summary>
-        protected java.util.List FieldOrigins =>
-            fieldOrigins ?? throw new InvalidOperationException("The statement has not been validated.");
-
-        /// <summary>
-        /// Runs the program, which is what chooses a plan.
-        /// </summary>
-        /// <param name="root"></param>
-        /// <returns></returns>
-        /// <remarks>
-        /// <c>Prepare.optimize</c>, less the materializations and lattices. See
-        /// <see cref="GetMaterializations"/>.
-        /// </remarks>
-        RelRoot Optimize(RelRoot root)
-        {
-            var planner = root.rel.getCluster().getPlanner();
-            planner.setExecutor(new RexExecutorImpl(context.getDataContext()));
-
-            var materializations = GetMaterializations();
-            var lattices = GetLattices();
-            var desiredTraits = GetDesiredRootTraitSet(root);
-
-            return root.withRel(GetProgram().run(planner, root.rel, desiredTraits, materializations, lattices));
-        }
-
-        /// <summary>
-        /// Returns the materialized views the planner may substitute.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         /// <remarks>
         /// Empty, where Calcite queries the <c>MaterializationService</c> and populates each through
         /// <c>CalcitePrepareImpl.populateMaterializations</c> — which builds a <c>CalciteMaterializer</c>, a
@@ -319,21 +164,18 @@ namespace Apache.Calcite.Extensions.Prepare
         /// It cannot be reached and cannot be written again, so a materialized view is not substituted here
         /// however the connection is configured.
         /// </remarks>
-        static java.util.List GetMaterializations()
+        protected override java.util.List GetMaterializations()
         {
             return com.google.common.collect.ImmutableList.of();
         }
 
-        /// <summary>
-        /// Returns the lattices the planner may use.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>Schemas.getLatticeEntries</c>, as Calcite's does. A lattice needs a star table to be useful and
         /// <see cref="GetMaterializations"/> supplies none, so this is empty in practice; it is read rather
         /// than assumed empty because that is a fact about the schema rather than about this class.
         /// </remarks>
-        java.util.List GetLattices()
+        protected override java.util.List GetLattices()
         {
             var entries = org.apache.calcite.schema.Schemas.getLatticeEntries(schema);
             var lattices = new java.util.ArrayList();
@@ -342,105 +184,57 @@ namespace Apache.Calcite.Extensions.Prepare
             {
                 var lattice = (CalciteSchema.LatticeEntry)i.next();
                 var starTable = lattice.getStarTable();
-                var starRelOptTable = RelOptTableImpl.create(catalogReader, starTable.getTable().getRowType(typeFactory), starTable, null);
+                var starRelOptTable = RelOptTableImpl.create(CatalogReader, starTable.getTable().getRowType(typeFactory), starTable, null);
                 lattices.add(new RelOptLattice(lattice.getLattice(), starRelOptTable));
             }
 
             return lattices;
         }
 
-        /// <summary>
-        /// Trims the fields no one reads.
-        /// </summary>
-        /// <param name="root"></param>
-        /// <returns></returns>
-        /// <remarks>
-        /// <c>Prepare.trimUnusedFields</c>. It builds a second converter because the trim runs with
-        /// <c>trimUnusedFields</c> decided per plan by <c>shouldTrim</c> rather than by the config the query
-        /// was converted with.
-        /// </remarks>
-        RelRoot TrimUnusedFields(RelRoot root)
-        {
-            var config = SqlToRelConverter.config()
-                .withTrimUnusedFields(ShouldTrim(root.rel))
-                .withExpand(((java.lang.Boolean)org.apache.calcite.prepare.Prepare.THREAD_EXPAND.get()).booleanValue())
-                .withInSubQueryThreshold(((java.lang.Integer)org.apache.calcite.prepare.Prepare.THREAD_INSUBQUERY_THRESHOLD.get()).intValue());
-
-            var converter = GetSqlToRelConverter(Validator, catalogReader, config);
-            var ordered = root.collation.getFieldCollations().isEmpty() == false;
-            var dml = SqlKind.DML.contains(root.kind);
-
-            return root.withRel(converter.trimUnusedFields(dml || ordered, root.rel));
-        }
-
-        /// <summary>
-        /// Returns whether a plan is worth trimming.
-        /// </summary>
-        /// <param name="rootRel"></param>
-        /// <returns></returns>
-        /// <remarks>
-        /// <c>Prepare.shouldTrim</c>: trimming a bare projection would leave nothing behind.
-        /// </remarks>
-        static bool ShouldTrim(RelNode rootRel)
-        {
-            return rootRel is not org.apache.calcite.rel.logical.LogicalProject;
-        }
-
-        /// <summary>
-        /// Flattens structured types.
-        /// </summary>
-        /// <param name="rootRel"></param>
-        /// <param name="restructure"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>CalcitePreparingStmt.flattenTypes</c>, which flattens only under a Spark handler and otherwise
         /// hands the plan back. This convention cannot use a Spark handler at all — it has no generated
         /// class to give one — so this is the branch Calcite takes without one.
         /// </remarks>
-        static RelNode FlattenTypes(RelNode rootRel, bool restructure)
+        protected override RelNode FlattenTypes(RelNode rootRel, bool restructure)
         {
             return rootRel;
         }
 
-        /// <summary>
-        /// Removes correlation from a plan.
-        /// </summary>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>CalcitePreparingStmt.decorrelate</c>. Reached only when the connection asks for it: a
         /// correlate is a node these conventions implement, and decorrelating removes the node they would
         /// have used.
         /// </remarks>
-        static RelNode Decorrelate(SqlToRelConverter sqlToRelConverter, SqlNode query, RelNode rootRel)
+        protected override RelNode Decorrelate(SqlToRelConverter sqlToRelConverter, SqlNode query, RelNode rootRel)
         {
             return sqlToRelConverter.decorrelate(query, rootRel);
         }
 
-        /// <summary>
-        /// Builds the converter from SQL to relational algebra.
-        /// </summary>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>CalcitePreparingStmt.getSqlToRelConverter</c>. This class is the view expander it is given.
         /// </remarks>
-        SqlToRelConverter GetSqlToRelConverter(SqlValidator validator, org.apache.calcite.prepare.Prepare.CatalogReader catalogReader, SqlToRelConverter.Config config)
+        protected override SqlToRelConverter GetSqlToRelConverter(SqlValidator validator, org.apache.calcite.prepare.Prepare.CatalogReader catalogReader, SqlToRelConverter.Config config)
         {
-            config = config.withTopDownGeneralDecorrelationEnabled(context.config().topDownGeneralDecorrelationEnabled());
+            config = config.withTopDownGeneralDecorrelationEnabled(Context.config().topDownGeneralDecorrelationEnabled());
 
             return new SqlToRelConverter(this, validator, catalogReader, cluster, convertletTable, config);
         }
 
-        /// <summary>
-        /// Renders a plan, for an <c>EXPLAIN</c>.
-        /// </summary>
+        /// <inheritdoc />
         /// <remarks>
         /// <c>Prepare.PreparedExplain</c>'s <c>getCode</c>, which is where the text is produced rather than
         /// at the point the statement is run.
         /// </remarks>
-        ClrPrepareResult CreatePreparedExplanation(
+        protected override ClrPrepareResult CreatePreparedExplanation(
             RelDataType? resultType,
             RelDataType parameterRowType,
             RelRoot? root,
-            org.apache.calcite.sql.SqlExplainFormat format,
-            org.apache.calcite.sql.SqlExplainLevel detailLevel)
+            SqlExplainFormat format,
+            SqlExplainLevel detailLevel)
         {
             var explanation = root == null
                 ? RelOptUtil.dumpType(requireResultType(resultType))
@@ -460,7 +254,7 @@ namespace Apache.Calcite.Extensions.Prepare
         /// </remarks>
         SqlValidator CreateSqlValidator(CalciteCatalogReader catalogReader, Func<SqlValidator.Config, SqlValidator.Config> configTransform)
         {
-            var connectionConfig = context.config();
+            var connectionConfig = Context.config();
             var opTab0 = (SqlOperatorTable)connectionConfig.fun((java.lang.Class)typeof(SqlOperatorTable), org.apache.calcite.sql.fun.SqlStdOperatorTable.instance());
 
             var list = new java.util.ArrayList();
@@ -501,7 +295,7 @@ namespace Apache.Calcite.Extensions.Prepare
                 throw new java.lang.RuntimeException("parse failed", e);
             }
 
-            var viewCatalogReader = catalogReader.withSchemaPath(schemaPath);
+            var viewCatalogReader = CatalogReader.withSchemaPath(schemaPath);
             var viewValidator = CreateSqlValidator(viewCatalogReader, c => c.withEmbeddedQuery(true));
             var config = SqlToRelConverter.config().withTrimUnusedFields(true);
             var sqlToRelConverter = GetSqlToRelConverter(viewValidator, viewCatalogReader, config);
